Moving the last set of tests.
